Making GEMINI query filter conditions more lenient (step 28)
I originally used the more strict/complex variation of the query conditions described in the tutorial. In order to analyse the effects of this in the final output I am now changing it to the standard leniency.

Previous conditions translated into “variants classified as somatic with a p-value <= 0.05, which haven’t been flagged as likely false-positives”. With these less complex conditions we retain all of the variants considered somatic variants by the variant caller.